tcpcryptd: disable ASM call to access stats about CPU cycles in ARM
There seems to be other APIs for accessing CPU cycling stats, but they
would need to be enabled in kernel mode.

Since this feature is only used in the profiler, I'll just disable it
for now and probably enable it again enabling the access for the stats
from the kernel space anytime soon.