linux: Initial update of the API changed in linux 3.8
This patch basically makes it compile against the version 3.8 of linux
with a patch that add tcpcrypt data structures to the TCP stack.

Header files were added, constant names changed structs updated
etc. It doesn't really guratantee that it will work, it is just the
first attempt to see it running.